Do not hide attachments when they are nested in a message part which is itself an attachment
I introduced a bug in commit c430463; the flags which control part display were
not properly trimmed when recursively descending into the nested MIME parts.
This had an unfortunate side effect that an attached message/rfc822 which itself
contained an attached PDF file would hide the PDF file from display.

This is a pretty critical problem -- hiding message parts which are there is
never a good thing :(.

v2: Changed the func name (to my taste) and simplified the bitwise operation.
Thanks to Thomas Lübking for review.

v3: Add an explicit enum close to the site where the original flags are defined
in a hope to catch mistakes when new flags are introduced in future. Thanks to
Thomas for this suggestion. I'm not adding it directly to the enum; I hope that
this way makes it clear that the result is *not* a flag in itself, just a
bitwise mask. Please comment within a review if you think this needs a change.

REVIEW: 112291