Add Qt private file qsystemlibrary_p.h which is needed by qstandardpaths_win.cpp
Not all Qt SDK packages contains qt private headers (specialy cross compiled
env) and compilation without qsystemlibrary_p.h failing for windows platform.

This is needed for mimetypes-qt4, the Qt4 port of Qt5's QMimeDatabase and
friends. The bacported QStandardPaths::writableLocation makes use of the
QSystemLibrary class from this header for resolving windows-specific paths. I
don't know whether QDesktopServices could be used for this purpose, but even if
that was possible, modifying the mimetypes library is not the thing to do
anyway.

REVIEW: 112373