Model: emit messageCountPossiblyChanged when periodically invalidating the message counts
The code which takes care of periodically refreshing the view of mailboxes emits
dataChanged() whenever it invalidates the cached message counts for a particular
mailbox. The systray code was, however, only listening to the
messageCountPossiblyChanged signal. During normal operation, this is not an
issue because the GUI's views would intercept the dataChanged and ask for the
actual data, which triggers an IMAP STATUS which in turn updates the message
counts and systray therefore notices that the numbers might have changed.
However, when the views are hidden for some reason, this process does not happen
and the systray icon is not updated.

CCBUG: 324309
REVIEW: 112474