tests: "fix" OS/2 unit tests
It looks that QTemporaryFile is broken on OS/2. I don't want to waste my time
debugging issues on a dead platform without being able to reproduce them myself,
so it's much easier to just QSKIP this test. Oh, and QSKIP is
source-incompatible between Qt4 and Qt5 for increased fun.

The other fix for IDLE shall be safe on other platforms; let's hope it fixes the
OS/2 problem. Probably related to some event loop timing or what not.

REVIEW: 112506
BUG: 324316