Update debian files
Reason for this patch is ability to create native package DEB on Debian/Ubuntu
systems with native tools for creating DEB packages (dpkg-buildpackage or
debuild wrapper).

Because we do not support old Ubuntu and Debian systems, debian package can be
updated to use debhelper version 8, which simplify debian/rules file.
Ubuntu precise has debhelper 9 and Debian squeeze has debhelper 8.

It is needed to test if this change does not break OBS.

REVIEW: 112600