Parser: simplify the number extraction
You might want to view this patch by `git show -w` to let it ignore whitespace
changes. This patch simplifies the code path (and as a side effect shall make it
very slightly faster to execute, but that's not the point of this patch -- that
function deals with QVariant after all).