Use the SettingsCategoryGuard everywhere
QSettings has a couple of similar methods, the
beginReadArray/beginWriteArray/endArray. These are *not* used by this patch; the
correct API would probably have to be built on iterators in order to be
reasonably useful. That's going to be much more involved, though, and therefore
outside of scope of this change -- patches welcome.

REVIEW: 113099