IMAP: detect UIDVALIDITY changes when they happen while a mailbox is already selected
No real-world IMAP server shall do this, ever, and expect the clients to work
well. That said, clients absolutely have to handle this reasonably. Given that
no server is known to do this (apart from an old version which does this at most
once for a user for reasons related to delayed mailbox allocation), it seems
acceptable to simply disconnect in this case.

Yes, it might be possible to reconnect to the mailbox. However, verifying that
all pending requests are killed reasonably is too much work for a condition that
no Trojita user will ever see, hopefully.

BUG: 324763
REVIEW: 113109