Assume that UID, FLAGS and MODSEQ are more common than other FETCH responses and move the comparison earlier
The code which handles the FETCH responses is scanning the passed data in order.
For each item in there, the code has to take a look at what "type of item" it
is, e.g. whether it's an UID, FLAGS, BODYSTRUCTURE or something else. This
jumping is currently performed linearly.

Some items are more frequent than the others. For example, almost every FETCH
response (save the unsolicited ones) contains the UID, so it makes sense to
check for that at first, and carry on in the order of how often these items are
expected to appear. On older IMAP servers, we're syncing FLAGS for the whole
mailbox all the time, so it makes sense to have FLAGS as the second option. The
modern servers are likely to return a MODSEQ with every FLAGS update (which are
less common, though), so it probably doesn't hurt to move these a bit up.

Using GCC 4.6.3 with -O3, the time spent in TreeItemMailbox::handleFetchResponse
is reduced from 12.63% to 11.79% of the total runtime.

REVIEW: 113121