Fix parsing to not recognize NILxxx as NIL
There was a long-standing bug in the parser which apparently treated any atoms
starting with NIL as NILs. The good thing about that was that it only happened
when the data were not transmitted as quoted strings or literals *and* the code
asked for the token via getAnything. It's still a pretty severe bug, though,
which could lead to stuff not showing up in Trojita for some reason.