tests: verifying RolePartId and actual item fetching
The code which handles part IDs was largely untested. This test makes it
possible to pass in the BODYSTRUCTURE and then iterate over the resulting tree
of indexes, checking that the "part ID" as used by the FETCH command and the
Trojita's internal code work reasonably.

Full rountrip via the FETCH command is performed.