Add accessor for isTopLevelMultipart
The IMAP model of the MIME tree is, let's say, a bit strange -- if a
message/rfc822 contains a multipart item as its root part, the child multipart
will *not* get assigned its own part number. The data are still available via
parent's .TEXT modifier, of course, but it causes all sorts of havoc in
Trojita's code. I do argue that the Trojita's model of MIME parts makes sense,
if only for providing a reasonable way for part rendering.