Do not attempt fetching special parts when they are not defined
The RFC3501 is clear on saying that the TEXT and HEADER modifiers are only valid
on message/rfc822 MIME types (either the root message, or any corresponding
nested message part).

The only problem with this is that we need to support the MIME modifier on most
of the other *parts* (but not on the top-level message). This means that we would
like to have most parts with just a single special column, all top-level
messages with two extra columns, and the TreeItemPartMultipartMessage with three
such modifiers. Unfortunately, these are conflicting requirements, so we have to
live with "holes" in the special columns where most parts return an invalid
index for columns #1 and #2, while they do contain valid data in #0 and #3. The
alternative would be to abandon enums which is even more evil.

And finally, if an IMAP part does not have a part number assigned (i.e. if it is
a top-level multipart), we cannot really ask for its "special parts" with
modifiers like HEADER, TEXT or MIME.