Graceful handling of runtime errors for IMAP attachment drag&drop
Originally, the code would happily assert crash when fed with malicious drop
data -- it simply assumed that the mailbox will always be there, that the
UIDVALIDITY has not changed, etc. This was always true for use within Trojita,
but it's unfortunately not the case when the input comes from outside.

This patch therefore lets the constructurs of IMAP attachment items to throw
exceptions, and the existing infrastructure is now used consistently so that an
item failing to add doesn't leak memory (hopefully).