Ignore duplicate QWebView::loadStarted events
It looks like certain messages like those from Mike Cardwell's Email Privacy
Tester cause WebKit to emit duplicate loadStarted() signals without the matching
loadFinished(). This in turn confuses our logic which mistakenly assumes that
for each loadStarted(), there will always be a matching loadFinished() as well.

And as a reminder, the reason why there's a set of items is because a single
message might consist of multiple individual parts and each of them is loaded
separately.

Thanks to Mike for reporting this.

REVIEW: 113308