Better checks of the part paths
Let's immediately return a nullptr in case the textual special part identifier
points to some garbage.

Also be sure to check that there are no extra chunks in our location after the
special parts is found. The old version would exit the loop unconditionally,
which is not the right thing to do. We want to return a nullptr on e.g. /TEXT/0,
not treat that as /TEXT.