IMAP: Implement fetching of raw part data without the CTE decoding
Certain applications, most notably support for signature verification, requires
access to raw, unprocessed part data as returned by the IMAP server. Trojita,
however, is happy to utilize the BINARY extension and stores the data in the
already decoded form.

This patch adds a new special tree item, the OFFSET_RAW_CONTENTS. When accessed,
a request for raw data is made, the result is stored in an appropriate cache
entry and the data are also made available to the "regular" part.