Imap: distinguish about fetching the processed and unprocessed data
This patch tries to make sure that the data are not kept twice in memory unless
asked to, and that the data in the on-disk cache are not duplicated.

One choice to make is what to do when we already know the decoded data and ask
for the raw data once again. The paranoid option would be to have a look at the
cached, already-decoded data and check whether what we have now matches. The
code will, however, simply leave the older data around and neither check nor
overwrite them with the locally decoded copy.