Remove processed data from cache when the raw data are stored
The idea is that both base64 and quoted-printable decoding is blazingly fast, so
it makes a lot of sense to conserve space by saving just a single copy on disk.

Note that no checks are made whether both copies of data are actually the same.
We do actively try to catch the server cheating here.