Do not set the $SubmitPending flag during the initial submission attempt
The $SubmitPending IMAP keywords  on its own is meant to indicate that "this
message needs to be sent out" [1]; MUAs are supposed to use both $Submitted and
$SubmitPending when the message is actively submitted, and just $Submitted when
the message was sent.

This is a serious bug -- a reasonable client could very well look at the Sent
mailbox and tell users "Hey, it looks like there's a ton of messages to be sent
in there which were never actually sent. Do you want me to send them now?". We
should have never set this flag without proper support for $Submitted in place,
but this misfeature has been around since mid-2012 :(.

Sorry to all who are affected by this, I haven't realized this implication.

[1] http://tools.ietf.org/html/rfc5550#section-5.10

REVIEW: 113396
CCBUG: 321381