Disable HTMLification of a too big plaintext
The current code for turning a plaintext into HTML is very slow on large
payload. Some of the time is spent in actual plaintext parsing and HTML
building, but even the WebKit is fed with the data twice instead of just once.

A proper fix would be to move this transformation outside of the GUI. An obvious
place is a QNetworkReply subclass with some help from the QNetworkAccessManager.
Once this is offloaded, we could even use a separate thread to not block the
GUI.

See http://thread.gmane.org/gmane.mail.trojita.general/915 for details.

REVIEW: 113421