Include real credits in the about box and make it prettier
The old about box only had a small icon with no real credits for all
contributors, just a simple link (which did not work because the proper QLabel
link delegation was not active).

The new version adds the LICENSE file with all contributors as a Qt resource and
uses the data at runtime after some transformation to make it better suited for
the about box presentation.

The layout of the UI got changed a bit to better accommodate (IMHO) the changed
amount of information we show.

REVIEW: 113526