QWebView: prevent repeated resizing of the contents
The call to constrainSize resulted in changing the size of the page contents
which led to constrainSize being called again. We were lucky that there wasn't
an infinite loop.

The reaction to page()->mainFrame()->contentsSizeChanged(QSize) is critical to
remain here because of e.g. the interactive resizing of JPEG images or for
proper loading of complex HTML mail which might refer to, say, other parts via
the CID mechanism.

REVIEW: 113538