tree: preallocate memory for all TreeItemMessage instances during the mailbox sync
There is no measurable performance change on my test setup, which is expected --
the amortized cost of reallocating is still O(1) per element, i.e. O(n) for the
mailbox as a whole, and pointer copying is dwarfed by the overhead of dynamic
allocation of the actual nodes and other useful work, but it's still good style
to preallocate what is known to be needed in future.

Thanks to Thomas Lübking for suggesting this.

REVIEW: 114124