GUI: use different icons for different window roles
I'm experimenting with a vertical DE panel which means that I only get an icon
for each window which makes it pretty hard to distinguish among them. Let's use
different icons for the main application window, for an individual message, for
the composer and for the headers/source view.

v2: Use Thomas' icons for the composer