GUI: make sure that all actions are reachable even when the menu is hidden
The action has to be added to a visible widget if we want the shortcuts to work.
The menu had an option for hiding for quite some time, but a hidden menu
apparently stops these shortcuts from working. As a workaround, we can easily
add the actions directly to the main widget.

v2: Move the add-action-to-widget code to the place where they're getting added
to the menu, and be sure it's unified into a single macro. Also remove extra
addAction calls where they are not needed.

REVIEW: 114183