tests: measure how quickly we can load flags from the cache and resync them
The scenario is as follows: the user has opened the mailbox before, but Trojita
was closed since that time. This means that the flags will be loaded from cache
when the user opens the mailbox next time. The goal here is including the
overhead of reading flags from cache in the resulting number. It takes roughly
3.8 - 3.9s on my T420s.