cache: Read the flags in one sweep instead of repeating it all the time
This commit introduces a new API for accessing the cached flags. Instead of
always looking up each and every flag by hand, we're now using a single sweep
through the table for all messages in a mailbox. This relies on the fact the the
items are already sorted by the UID, so the sqlite's native record stepping
works to our advantage. Unfortunately, it might well happen that the cache can
contain additional extra items (the cache on my laptop certainly does), so
there's still that fallback for accessing each item individually. This might
change in future -- perhaps the data shall be squashed to  just a single row in
the table at the cost of maybe more expensive updates.

I guess I have more measurements to make here.

Some of the tests are now a bit more noisy due to these checks; this is
something to be solved later on.