IMAP: fix a bug when opening folders when offline after a STATUS arrived
By definition, the mailboxes are synchronized when they are opened. When all
whic his needed is just the number of e.g. unread/total messages, a lightweight
variant of syncing is performed through the STATUS command. The STATUS data are
displayed in the GUI, but they are not cached. Unfortunately, they also managed
to interfere with opening the mailbox from a locally cached copy when offline.

This change makes sure that the only "cache consistency check" which is
performed happens on top of the cached data, i.e. the UID mapping and the
synchronization state. What has in the meanwhile arrived via STATUS is
completely irrelevant and should not affect the synchronization process (we do
not even know whether some messages were deleted, so it is impossible to show
e.g. a couple (status_exists - old_exists) empty placeholder messages, etc).

I guess it's very handy that I travel on train every now and then :).

REVIEW: 114225