IMAP: UNSELECT: only ignore untagged responses
The intention of that code was to only ingore responses which typically arrive
during the opening of the "non-existing" mailbox. The tagged responses which
also happen to be marked with some particular response code shall fall through
to the handler of the respective tag, otherwise we might eat something which is
important.