Remove explicit "response kind" tracking which is already done at the class level
The idea is that each class shall correspond to a single response, more or less.
In practice, we use some of the classes for multiple responses which are in fact
very similar. This has led to some duplication where the AbstractResponse::kind
was often shadowed by the same member defined in the child class, which was not
funny to see in backtraces.

This patch removes the AbstractResponse::kind altogether. The member is onstead
only used in those subclasses which actually make use of it.