GUI: move the "mark as reply" selector to the row of buttons
There's a GCI task for changing various replying modes from within the composer
(think reply-to-all vs. reply-privately). That change needs a button to be added
somewhere, so it makes sense to use a button for the feature which Thomas added
yesterday. Besides, it seems to me that it is a bit easier to discover this way.

This is best viewed as a diff from commit
944a49e33562bd508a2af6e0dcfc938ba2769f1b; that way it is clearer what has
changed compared to the standalone-checkbox approach.

v2 changes:
- Hide the button altogether when not replying instead of just disabling it.
- Pretty formatting of the tooltips and sanitization of HTML input
- Better captions

Thanks to Thomas Lübking for his comments.

v3 changes:
- Enable quick toggling via a default action, without expanding the menu
- Fix restoring of v2 drafts

v4 changes:
- Removed one superfluous QAction::trigger
- Better identifiers for the slots for updating & processing the quick toggle
action
- Changed the caption of the threaded reply marking action
- Some inline comments

REVIEW: 114815