IMAP: do not assert crash on new arrivals during an in-progress FETCH
Some time ago the code was changed so that it reset the TreeItemMsgList's
fetching state back into LOADING when new arrivals were detected; this was
designed to make sure that the sync state and UIDs are always saved
consistently. Unfortunately, some other parts of the code tried ty depend on the
TreeItemMsgList already being fetched(), which is now no longer true.

Simply removing the assert works. The test case tries to check for this scenario
more aggressively, and also makes sure that we support the same not only for
part fetching, but for message metadata fetches as well.

BUG: 329757
REVIEW: 115009