net: Connect to the network even when the QtBearer is missing
On Gentoo, the dev-qt/qtbearer package was previsouly not properly added to
RDEPEND, which resulted in people not having it available. This resulted in
Trojita asking Qt to open a network session whose configuration was not actually
valid, which failed (no surprise on that one). Unfortunately, this error was not
user-visible at all, unless they opened the debug log themselves (and we already
know that nobody does that -- I wouldn't do it, either).

This patch adds a speculative shortcut to the network session management -- if
Qt claims that there is no configured network profile, we will assume that it is
wrong and try to connect anyway.

In addition, errors related to the network session establishment are now
communicated properly to the user.

Thanks to Stefan de Konink <stefan@konink.de> (skinkie on IRC) for a bugreport.

REVIEW: 115131