Add skeleton support for tracking the connection method "properly"
There's a couple of problems, as usual. First of all, the enum really has to
live outside of ImapAccess in orderd to make this useful to other classes (like
the SMTP thing) in future. However, this use case is apparently not supported on
with QML where the MOC absolutely needs the enum to live in a class derived from
QObject.

I'm fed up with this, so in the end I decided to just let the
backward-compatibility functions sslMode/setSslMode live for the time being, and
focus my effort somewhere else than towards fighting the moc abomination.

BURN IN HELL, MOC!