Composer: handle replying to messages with malformed Message-Id
It's a question what would be the best place to put this check at. Moving it all
the way to the MessageComposer is tempting, but then the GUI could still show
invalid values for the "this will be marked as a reply to ..." indicators.
Moving this into the middle layers like the composer widget itself sounds just
about wrong.

If this was strictly about filtering invalid values from e.g. References, it
might even make sense to push this all the way to Message.cpp, but we have to
handle also messages whose Message-Id is simply missing -- a situaiton which is
forbidden, but could happen in real word anyway. One cannot just invent a random
Message-Id, of course.

REVIEW: 115939