RFC5322: Support stray CR and LF wherever a CRLF pair is expected in the whitespace
Thanks to Arnt for testing Trojita against the aox. It turned out that the IMAP
server sent us a Message-Id which started as "\n <actual-data>", which is not
allowed as per RFC5322.

This is a change which I do not really want to make; even the obsolete grammar
does not include this relaxed rule. However, while we were so far lucky with
strictly conformant IMAP servers, there are certainly cases out there where a
simple LF is all what we will get, and it makes a lot of sense to be able to
work under these circumstances as well.

The actual header separator is *not* affected by this; we still expect our
headers to be CRLF-terminated. We'll see if this rule needs relaxing in future.

REVIEW: 115940