Fix the version string once again
OK, the rules are the following:
- A release "x.y" should have exactly "x.y" here, nothing more, nothing less
- Right after x.y is released, the string shall be changed back to "vx.y-git"

The idea is that *releases* always contain just numbers, while everything else
(and therefore something which is work-in-progress) should contain the "v"
prefix and "-git" suffix.

I hope I'll be able to make this happen in future; so far, the 0.3.96 was
different, and I just forgot to add "v" back in after the 0.4 was released. Meh.