IMAP: fix deleting of currently selected mailboxes
All deletion attempts are now scheduled via the KeepMailboxOpenTask. If the
mailbox being deleted is currently selected, it will be destructively closed
(and a random subset of messages removed, see RFC 3501) by sending of the CLOSE
command. Afterwards, the actual deletion happens as usual.

There's a couple of corner cases, hence the number of test cases. Hope this is
enough.