IMAP: do not request STATUS of mailboxes which are going to be replaced
The LIST processing will replace the cached mailboxes with completely new
instances, so it doesn't make sense to issue a STATUS for something which is
going to be replaced anyway.

When LIST-STATUS is present, this makes even more sense -- even if the TreeItem
nodes were not replaced, the STATUS response would still arrive, so there is
little point in requesting it explicitly. But that's just a remark for future;
the TreeItemMailbox nodes are for now being replaced when a LIST response
arrives.