Return previously known values when asked for message counts
When a message count is invalidated for some reason (perhaps because the
periodic timer has run), the GUI would not show the number of message at all.
This is suboptimal; in the name of getting a more accurate result in future,
what we get is that the actual value disappears for a while. This change makes
sure that we never forget the old data, even when asked for the new ones.