Do not claim the numbers are fresh when loading from cache
When a list of child mailboxes is loaded from the cache, it used to be shown as
grayed-out to make it obvious that the value might very well be stale. The
recent changes made them appear normal, which might be confusing.

We're cheating in this commit -- the value is marked as loading, so it will be
again shown as grayed-out, but there is no matching network request. Instead,
the code knows that the list of mailboxes will be replaced when the LIST
response arrives.