IMAP: save the number of unread/... messages during every mailbox sync
The sync state contains a ton of properties which are not critical for mailbox
updates, but are used by the GUI when showing states of the mailboxes. As it
happens, these numbers were not getting updated at the time the mailbox state
itself was recalculated and pushed to cache.

What should be done of top of this is saving the numbers whenever the STATUS
response arrives. Any change doing that must make sure that it does not clobber
the important SyncState members which are used for synchronization, though.