IMAP: store the results of STATUS queries into the cache
See comments inline for why it's important to *not* save EXISTS, UIDNEXT and
other sync-related data. We're saving just those bits which are useful for the
user, but *not* critical for the sync.

In future, this might be extended to also save the total number of messages, but
if this ever happens, it MUST NOT go into the EXISTS field or mailbox syncing
will be utterly broken.